% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio.R
\name{audio_load}
\alias{audio_load}
\title{Load an audio form a WAV file}
\usage{
audio_load(audio_path, downsample = NULL, ...)
}
\arguments{
\item{audio_path}{path to the WAVE file}

\item{downsample}{Sampling rate to which the output will be downsampled}

\item{...}{
  Arguments passed on to \code{\link[audio:wave]{audio::load.wave}}
  \describe{
    \item{\code{where}}{file name of the file to load from or save to}
  }}
}
\description{
Load an audio form a WAV file
}
