% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audio.R
\name{audio_spectrogram}
\alias{audio_spectrogram}
\title{Load an audio form a WAV file}
\usage{
audio_spectrogram(x, ...)
}
\arguments{
\item{x}{an audio object}

\item{...}{
  Arguments passed on to \code{\link[seewave:spectro]{seewave::spectro}}
  \describe{
    \item{\code{wave}}{an R object.}
    \item{\code{f}}{sampling frequency of \code{wave} (in Hz). Does not need to
    be specified if embedded in \code{wave}.}
    \item{\code{channel}}{channel of the R object, by default left channel (1).}
    \item{\code{wl}}{window length for the analysis
    (even number of points) (by default = 512).}
    \item{\code{wn}}{window name, see \code{\link[seewave]{ftwindow}} (by default
    \code{"hanning"}).}
    \item{\code{zp}}{zero-padding (even number of points), see \code{Details}.}
    \item{\code{ovlp}}{overlap between two successive windows (in \%).}
    \item{\code{noisereduction}}{a numeric vector of length 1, if \code{1} a
  noise reduction is applied along the rows of the spectrogram, if
  \code{2} a noise reduction applied along the columns. See \code{Details}.}
    \item{\code{fastdisp}}{faster graphic display for long \code{wave}. The
  spectrogram/oscillogram is displayed/saved faster in the graphic device/
  graphic file when set to \code{TRUE}, with a cost on graphical resolution.}
    \item{\code{complex}}{if \code{TRUE} the STFT will be returned as complex
  numbers.}
    \item{\code{norm}}{if \code{TRUE} the STFT is normalised (i. e. scaled) by its maximum.}
    \item{\code{correction}}{a character vector of length 1 to apply an
  amplitude ("amplitude") or an energy ("energy") correction
  to each FT window. This argument is useful only when one wish to obtain
  absolute values that is when \code{norm=FALSE}.
  By default no correction is applied ("none").}
    \item{\code{fftw}}{if \code{TRUE} calls the function \code{FFT} of the
  library \code{fftw}. See Notes.}
    \item{\code{dB}}{a character string specifying the type dB to return: "max0"
    (default) for a maximum dB value at 0, "A", "B", "C", "D", and "ITU" for
  common dB weights. If set to \code{NULL}, then a linear scale is used.}
    \item{\code{dBref}}{a dB reference value. \code{NULL} by default
    but should be set to 2*10e-5 for a 20 microPa reference.}
    \item{\code{plot}}{logical, if \code{TRUE} plots the spectrogram
    (by default \code{TRUE}).}
    \item{\code{flog}}{a logical to plot the frequency on a logarithmic scale.}
    \item{\code{grid}}{logical, if \code{TRUE} plots a y-axis grid
    (by default \code{TRUE}).}
    \item{\code{osc}}{logical, if \code{TRUE} plots an oscillogram beneath
    the spectrogram (by default \code{FALSE}).}
    \item{\code{scale}}{logical, if \code{TRUE} plots a dB colour scale on the right
    side of the spectrogram (by default \code{TRUE}).}
    \item{\code{cont}}{logical, if \code{TRUE} overplots contour lines on the spectrogram
    (by default \code{FALSE}).}
    \item{\code{collevels}}{a set of levels which are used to partition
    the amplitude range of the spectrogram (in dB).}
    \item{\code{palette}}{a color palette function to be used to assign colors in
    the plot, see \code{Details}.}
    \item{\code{contlevels}}{a set of levels which are used to partition
    the amplitude range for contour overplot (in dB).}
    \item{\code{colcont}}{colour for \code{cont} plotting.}
    \item{\code{colbg}}{background colour.}
    \item{\code{colgrid}}{colour for \code{grid} plotting.}
    \item{\code{colaxis}}{color of the axes.}
    \item{\code{collab}}{color of the labels.}
    \item{\code{cexlab}}{size of the labels.}
    \item{\code{cexaxis}}{size of the axes.}
    \item{\code{tlab}}{label of the time axis.}
    \item{\code{flab}}{label of the frequency axis.}
    \item{\code{alab}}{label of the amplitude axis.}
    \item{\code{scalelab}}{amplitude scale label.}
    \item{\code{main}}{label of the main title.}
    \item{\code{scalefontlab}}{font of the amplitude scale label.}
    \item{\code{scalecexlab}}{cex of the amplitude scale label.}
    \item{\code{axisX}}{logical, if \code{TRUE} plots time X-axis (by default \code{TRUE}).}
    \item{\code{axisY}}{logical, if \code{TRUE} plots frequency Y-axis (by default \code{TRUE}).}
    \item{\code{tlim}}{modifications of the time X-axis limits.}
    \item{\code{trel}}{time X-axis with a relative scale when \code{tlim} is not null,
    \emph{i.e.} relative to \code{wave}.}
    \item{\code{flim}}{modifications of the frequency Y-axis limits (in kHz).}
    \item{\code{flimd}}{dynamic modifications of the frequency Y-axis limits. New \code{wl}
    and \code{ovlp} arguments are applied to increase time/frequency resolution.}
    \item{\code{widths}}{a vector of length 2 to control the relative widths of columns on
  the device when \code{scale} is \code{TRUE}.}
    \item{\code{heights}}{a vector of length 2 to control the relative heights of rows on
    the device when \code{osc} is \code{TRUE}.}
    \item{\code{oma}}{a vector of length 4 to control the size of outer margins
    when either \code{scale} or \code{osc} is \code{TRUE}.}
    \item{\code{listen}}{if \code{TRUE} the sound is played back (by default
  \code{FALSE}).}
  }}
}
\description{
Load an audio form a WAV file
}
