% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{create_project_folders}
\alias{create_project_folders}
\title{Create project folders}
\usage{
create_project_folders(folders = NULL)
}
\arguments{
\item{x}{A dataframe}
}
\description{
Create project folders
}
