% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{flatten_columns}
\alias{flatten_columns}
\title{Transform any list column in a dataframe to collapsed character vector}
\usage{
flatten_columns(x)
}
\arguments{
\item{x}{A dataframe}
}
\description{
Transform any list column in a dataframe to collapsed character vector
}
