% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corpora.R
\name{get_childes_frequencies}
\alias{get_childes_frequencies}
\title{Get CHILDES lexical frequencies}
\usage{
get_childes_frequencies(collection = "Eng-NA", age_range = c(10, 36), ...)
}
\arguments{
\item{collection}{CHILDES corpora from where to fetch transcriptions. Takes "Eng-NA" (North American English by default). See \href{https://childes.talkbank.org/access/}{CHILDES Index to corpora} to see options}

\item{age_range}{Numeric vector of lenght two indicating the minimum and maximum age range of interest for which to comoute lexical frequencies in the CHILDES corpora. Frequencies will be summarised across this age range using the mean}

\item{...}{Additional arguments passed to \code{childesr::get_types}}
}
\description{
Get CHILDES lexical frequencies
}
