% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/participants.R
\name{months_to_years}
\alias{months_to_years}
\title{Transform months to years and months}
\usage{
months_to_years(x, .sep = ";")
}
\arguments{
\item{x}{Age in months}

\item{.sep}{Separator between years and months, ';' by default}
}
\description{
Transform months to years and months
}
