% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/participants.R
\name{prop_adj}
\alias{prop_adj}
\title{Adjusted estimated proportion of successes from Gelman, Hill & Vehtari (2020)}
\usage{
prop_adj(y, n)
}
\arguments{
\item{y}{Number of successes}

\item{n}{Number of trials}
}
\description{
Adjusted estimated proportion of successes from Gelman, Hill & Vehtari (2020)
}
