% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/participants.R
\name{prop_adj_ci}
\alias{prop_adj_ci}
\title{Confidence interval of the adjusted proportion of successes from Gelman, Hill & Vehtari (2020)}
\usage{
prop_adj_ci(y, n, .width = 0.95, limit)
}
\arguments{
\item{y}{Number of successes}

\item{n}{Number of trials}

\item{.width}{Width of the confidence interval (0.95 by default)}
}
\description{
Confidence interval of the adjusted proportion of successes from Gelman, Hill & Vehtari (2020)
}
