% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/participants.R
\name{prop_adj_se}
\alias{prop_adj_se}
\title{Standard error of the adjusted proportion of successes from Gelman, Hill & Vehtari (2020)}
\usage{
prop_adj_se(y, n)
}
\arguments{
\item{y}{Number of successes}

\item{n}{Number of trials}
}
\description{
Standard error of the adjusted proportion of successes from Gelman, Hill & Vehtari (2020)
}
