% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{remove_nul}
\alias{remove_nul}
\title{Remove nul file that get created sometimes when knitting Rmarkdown files to PDF}
\usage{
remove_nul()
}
\description{
Remove nul file that get created sometimes when knitting Rmarkdown files to PDF
}
