% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{rescale_variable}
\alias{rescale_variable}
\title{Rescale standardised variable}
\usage{
rescale_variable(x, mean, sd)
}
\arguments{
\item{x}{Numeric vector to be rescaled}

\item{mean}{Numeric value indicating the mean of the original vector}

\item{sd}{Numeric value indicating the standard deviation of the original vector}
}
\description{
Rescale standardised variable
}
