% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saptools-package.R
\docType{package}
\name{saptools-package}
\alias{saptools}
\alias{saptools-package}
\title{saptools: An R package for the SAP group}
\description{
Custom functions and scripts for the Speech Acquisition and Production research group at the Center for Brain and Cognition (Universitat Pompeu Fabra). An opinionated project structure and tidyverse-based workflow is suggested. This package also contains wrappers for setting up {targets} and {renv} to manage more complex workflows and package depencies, and Quarto for writing reports, lab notes, and manuscripts. Apache Arrow is use for loading and saving data in CSV, Parquet, and Arrow Database formats via the {arrow} R package.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
