% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\name{save_files}
\alias{save_files}
\title{Save an R object as data or results as Arrow Parquet, CSV, or RDS files}
\usage{
save_files(
  x,
  folder,
  file_name = deparse(substitute(x)),
  formats = c("parquet", "csv", "rds"),
  .sep = "/"
)
}
\arguments{
\item{x}{A tabular R object}

\item{folder}{Folder in which to write the file. If the resulting path does not exist, a new directory will be generated.}

\item{formats}{Formats in which to write the files. Must be at least one of 'parquet', 'csv', or 'rds'}

\item{.sep}{Path separator, takes '/' by default}
}
\description{
Save an R object as data or results as Arrow Parquet, CSV, or RDS files
}
