% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{seq_range}
\alias{seq_range}
\title{Generate a regular sequence of \code{n} elements in the range of a numeric vector \code{x}}
\usage{
seq_range(x, n)
}
\arguments{
\item{x}{Numeric vector}

\item{n}{Length of the vector to be generated}
}
\description{
Generate a regular sequence of \code{n} elements in the range of a numeric vector \code{x}
}
