% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{unmake}
\alias{unmake}
\title{Remove targets products}
\usage{
unmake(keep_fits = TRUE)
}
\arguments{
\item{keep_fits}{Should RDS object associated with model fits be kept?}
}
\description{
Remove targets products
}
